refactor(tagparser)
