+++ src/ContentFactory/PageFactory.cs
@@ -2,10 +2,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using form_builder.Extensions;
-using form_builder.Factories.TagParser;
 using form_builder.Helpers.PageHelpers;
 using form_builder.Models;
 using form_builder.Providers.StorageProvider;
+using form_builder.TagParser;
 using form_builder.ViewModels;
 using Newtonsoft.Json;
