+++ src/TagParsers/ITagParser.cs
@@ -1,7 +1,7 @@
 using System.Text.RegularExpressions;
 using form_builder.Models;
 
-namespace form_builder.Factories.TagParser
+namespace form_builder.TagParser
 {
     public interface ITagParser
     {
