+++ src/Utils/ServiceCollectionExtensions/ServiceCollectionExtensions.cs
@@ -6,7 +6,6 @@ using Amazon.SimpleEmail;
 using form_builder.Configuration;
 using form_builder.ContentFactory;
 using form_builder.Factories.Schema;
-using form_builder.Factories.TagParser;
 using form_builder.Factories.Transform.Lookups;
 using form_builder.Factories.Transform.ReusableElements;
 using form_builder.Gateways;
@@ -40,6 +39,7 @@ using form_builder.Services.PayService;
 using form_builder.Services.RetrieveExternalDataService;
 using form_builder.Services.StreetService;
 using form_builder.Services.SubmitService;
+using form_builder.TagParser;
 using form_builder.Validators;
 using form_builder.Workflows;
 using form_builder.Workflows.ActionsWorkflow;
